[SYCL-MLIR] Implement sycl-method-to-sycl-call pass
Implement pass to transform SYCLMethodOpInterface instances to
SYCLCallOps.
