+++ mlir-sycl/lib/Transforms/CMakeLists.txt
@@ -1,0 +1,14 @@
+add_mlir_library(MLIRSYCLTransforms
+  SYCLMethodToSYCLCall.cpp
+
+  ADDITIONAL_HEADER_DIRS
+  ${MLIR_MAIN_INCLUDE_DIR}/mlir/Transforms
+
+  DEPENDS
+  MLIRSYCLTransformsIncGen
+
+  LINK_LIBS PUBLIC
+  MLIRPass
+  MLIRTransformUtils
+  MLIRSYCLDialect
+  )
