+++ mlir-sycl/lib/Transforms/PassDetail.h
@@ -1,0 +1,24 @@
+//===- PassDetail.h - Transforms Pass class details -------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef TRANSFORMS_PASSDETAIL_H_
+#define TRANSFORMS_PASSDETAIL_H_
+
+#include "mlir/Dialect/SYCL/IR/SYCLOpInterfaces.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Pass/PassManager.h"
+#include "mlir/Transforms/Passes.h"
+
+namespace mlir {
+namespace sycl {
+#define GEN_PASS_CLASSES
+#include "mlir/Dialect/SYCL/Transforms/Passes.h.inc"
+} // namespace sycl
+} // namespace mlir
+
+#endif // TRANSFORMS_PASSDETAIL_H_
