+++ polygeist/lib/polygeist/Passes/ConvertToLLVMABI.cpp
@@ -178,94 +178,6 @@ public:
   }
 };
 
-static mlir::Value castToBaseType(PatternRewriter &Rewriter, mlir::Location Loc,
-                                  mlir::Value Original,
-                                  mlir::MemRefType BaseType) {
-  const auto ThisType = Original.getType().cast<MemRefType>();
-  const llvm::ArrayRef<int64_t> TargetShape = BaseType.getShape();
-  const mlir::Type TargetElementType = BaseType.getElementType();
-  const unsigned TargetMemSpace = BaseType.getMemorySpaceAsInt();
-
-  assert(TargetShape == ThisType.getShape() &&
-         "Shape should not change when casting to base class for a member "
-         "function call.");
-  assert(TargetMemSpace == ThisType.getMemorySpaceAsInt() &&
-         "Memory space of the `this` argument should be preserved when "
-         "creating a SYCLMethodOp instance.");
-
-  // The element type will always change here.
-  mlir::Value Cast = Rewriter.create<sycl::SYCLCastOp>(
-      Loc,
-      MemRefType::get(TargetShape, TargetElementType, {},
-                      ThisType.getMemorySpaceAsInt()),
-      Original);
-
-  LLVM_DEBUG(llvm::dbgs() << "  Cast inserted: " << Cast << "\n");
-
-  return Cast;
-}
-
-static LogicalResult convertMethod(sycl::SYCLMethodOpInterface method,
-                                   PatternRewriter &rewriter) {
-  LLVM_DEBUG(llvm::dbgs() << "ConvertToLLVMABIPass: SYCLMethodOpLowering: ";
-             method.dump(); llvm::dbgs() << "\n");
-
-  SmallVector<mlir::Value> Args(method->getOperands());
-
-  const auto BaseType = method.getBaseType().cast<MemRefType>();
-  if (BaseType != Args[0].getType().cast<MemRefType>())
-    Args[0] = castToBaseType(rewriter, method->getLoc(), Args[0], BaseType);
-
-  const auto ResTyOrNone = [=]() -> llvm::Optional<mlir::Type> {
-    const auto ResTys = method->getResultTypes();
-    if (ResTys.empty())
-      return llvm::None;
-    assert(ResTys.size() == 1 && "Returning multiple values is not allowed in "
-                                 "SYCLMethodOpInterface instances");
-    return ResTys[0];
-  };
-
-  auto CallOp = rewriter.replaceOpWithNewOp<sycl::SYCLCallOp>(
-      method, ResTyOrNone(), method.getTypeName(), method.getFunctionName(),
-      method.getMangledFunctionName(), Args);
-
-  LLVM_DEBUG(llvm::dbgs() << "  Converted to: " << CallOp << "\n");
-
-  return convertOperation(*method, rewriter);
-}
-
-template <typename T, typename = std::enable_if_t<sycl::isSYCLMethod<T>::value>>
-class SYCLMethodOpLowering final : public OpRewritePattern<T> {
-public:
-  using OpRewritePattern<T>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(T op, PatternRewriter &rewriter) const final {
-    return convertMethod(op, rewriter);
-  }
-};
-
-// If the operation is a method, add the SYCLMethod pattern for that
-// operation.
-template <typename T>
-static typename std::enable_if_t<mlir::sycl::isSYCLMethod<T>::value>
-addSYCLMethodPattern(RewritePatternSet &patterns, ConversionTarget &target,
-                     MLIRContext *ctx) {
-  patterns.add<SYCLMethodOpLowering<T>>(ctx);
-  target.addIllegalOp<T>();
-}
-
-// If the operation is not a method, do nothing.
-template <typename T>
-static typename std::enable_if_t<!mlir::sycl::isSYCLMethod<T>::value>
-addSYCLMethodPattern(RewritePatternSet &, ConversionTarget &, MLIRContext *) {}
-
-template <typename... Args>
-static void addSYCLMethodPatterns(RewritePatternSet &patterns,
-                                  ConversionTarget &target, MLIRContext *ctx) {
-  (void)std::initializer_list<int>{
-      0, (addSYCLMethodPattern<Args>(patterns, target, ctx), 0)...};
-}
-
 struct ConvertToLLVMABIPass final
     : public mlir::polygeist::ConvertToLLVMABIBase<ConvertToLLVMABIPass> {
   void runOnOperation() override {
@@ -281,12 +193,6 @@ struct ConvertToLLVMABIPass final
     patterns.add<SYCLConstructorLowering>(&getContext());
 
     ConversionTarget target(getContext());
-
-    // Add SYCLMethodOp conversion patterns and illegalize these operations.
-    addSYCLMethodPatterns<
-#define GET_OP_LIST
-#include "mlir/Dialect/SYCL/IR/SYCLOps.cpp.inc"
-        >(patterns, target, &getContext());
 
     auto checkFuncTy = [](FunctionType funcTy) {
       bool hasMemRef = llvm::any_of(
