+++ polygeist/tools/cgeist/CMakeLists.txt
@@ -68,6 +68,7 @@ target_link_libraries(cgeist PRIVATE
   MLIROpenMPToLLVM
   MLIROpenMPToLLVMIRTranslation
   MLIRSYCLToLLVM
+  MLIRSYCLTransforms
   
   LLVMSYCLLowerIR
